Moved generator into its own compileer-package, and renamed it
